+++ src/app/components/profile/UserAvatar.tsx
